refactor interface in 3.services
