+++ course3/3.clean_architecture/3.services/task3.3.3.1/internal/controller/conrtoller.go
@@ -21,6 +21,15 @@ type GeoController struct {
 	service   GeoServicer
 }
 
+type Responder interface {
+	OutputJSON(w http.ResponseWriter, data any)
+
+	ErrorUnAuthorized(w http.ResponseWriter, err error)
+	ErrorBadRequest(w http.ResponseWriter, err error)
+	ErrorInternal(w http.ResponseWriter, err error)
+	ErrorForbidden(w http.ResponseWriter, err error)
+}
+
 func NewGeoController(responder Responder, serv GeoServicer) *GeoController {
 	return &GeoController{
 		responder: responder,
