+++ course3/3.clean_architecture/3.services/task3.3.3.1/internal/controller/responder.go
@@ -8,15 +8,6 @@ import (
 	"net/http"
 )
 
-type Responder interface {
-	OutputJSON(w http.ResponseWriter, data any)
-
-	ErrorUnAuthorized(w http.ResponseWriter, err error)
-	ErrorBadRequest(w http.ResponseWriter, err error)
-	ErrorInternal(w http.ResponseWriter, err error)
-	ErrorForbidden(w http.ResponseWriter, err error)
-}
-
 type Response struct {
 	Success bool   `json:"success"`
 	Data    any    `json:"data"`
